api/requirements.txt
psycopg2-binary
fastapi
uvicorn
docker
requests
pycryptodome
lark-oapi
nest-asyncio
